requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
asttokens==3.0.1
click==8.3.1
comm==0.2.3
contourpy==1.3.3
cycler==0.12.1
decorator==5.2.1
drjit==1.2.0
executing==2.2.1
fastapi==0.124.4
fonttools==4.61.1
h11==0.16.0
idna==3.11
importlib_resources==6.5.2
ipydatawidgets==4.3.5
ipython==9.8.0
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.8
jedi==0.19.2
jupyterlab_widgets==3.0.16
kiwisolver==1.4.9
matplotlib==3.10.8
matplotlib-inline==0.2.1
mitsuba==3.7.1
numpy==2.3.5
packaging==25.0
parso==0.8.5
pexpect==4.9.0
pillow==12.0.0
prompt_toolkit==3.0.52
ptyprocess==0.7.0
pure_eval==0.2.3
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pyparsing==3.2.5
python-dateutil==2.9.0.post0
pythreejs==2.4.2
scipy==1.16.3
sionna-rt==1.2.1
six==1.17.0
stack-data==0.6.3
starlette==0.50.0
traitlets==5.14.3
traittypes==0.2.3
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.38.0
wcwidth==0.2.14
widgetsnbextension==4.0.15
pyproj==3.7.2
